requirements.txt
asgiref==3.7.2
certifi==2023.5.7
Django==4.2.2
django-haystack==3.2.1
djangorestframework==3.14.0
elasticsearch==7.17.9
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
